[oarsh] add srun like feature to oarsh: exec on subset of cgroup
ex:
Create job with 2 CPUs.
> oarsub -l /host=2/cpu=2 ...

Get the cpuset ids of the first cpu
> oarprint cpu -P host,cpuset
node1 0,2,4,6
node1 1,3,5,7
node2 0,2,4,6
node2 1,3,5,7

Execute a commande on the first cpu of node1
> export OAR_USER_CPUSET=$(oarprint cpu -P host,cpuset | grep node1 | head -n1 | cut -d\  -f2)
> oarsh node1

Then on node1:
> cat /dev/cpuset/oar/user_jobid/cpuset.cpus
0-7
> cat /proc/self/cpuset
/oar/user_jobid/0,2,4,6
> cat /dev/cpuset/oar/user_jobid/0,2,4,6/cpuset.cpus
0,2,4,6

And works as well for GPUs: given a job on a host with 4 GPUs
> OAR_USER_GPUDEVICE=1,2 oarsh node1
only gpu 1 and 2 (/dev/nvidia1 and /dev/nvidia2) will show up in
nvidia-smi, not gpu 0 and 3.

Of course, OAR_USER_CPUSET and OAR_USER_GPUDEVICE cannot be used to gain
access to resources which are not part of the job (thanks to the cgroup
hierarchy)

+++ sources/core/tools/job_resource_manager_cgroups.pl
@@ -450,11 +450,15 @@ EOF
     # Clean cpuset on this node
     if (defined($Cpuset_path_job)){
         system('echo THAWED > '.$Cgroup_directory_collection_links.'/freezer/'.$Cpuset_path_job.'/freezer.state
-                PROCESSES=$(cat '.$Cgroup_directory_collection_links.'/cpuset/'.$Cpuset_path_job.'/tasks)
-                while [ "$PROCESSES" != "" ]
-                do
-                    oardodo kill -9 $PROCESSES > /dev/null 2>&1
-                    PROCESSES=$(cat '.$Cgroup_directory_collection_links.'/cpuset/'.$Cpuset_path_job.'/tasks)
+                for d in '.$Cgroup_directory_collection_links.'/cpuset/'.$Cpuset_path_job.'/* '.$Cgroup_directory_collection_links.'/cpuset/'.$Cpuset_path_job.'; do
+                  if [ -d $d ]; then
+                    PROCESSES=$(cat $d/tasks)
+                    while [ "$PROCESSES" != "" ]
+                    do
+                        oardodo kill -9 $PROCESSES > /dev/null 2>&1
+                        PROCESSES=$(cat $d/tasks)
+                    done
+                  fi
                 done');
         
         # Locking around the cleanup of the cpuset for that user, to prevent a creation to occure at the same time
@@ -464,15 +468,20 @@ EOF
             if (system('if [ -w '.$Cgroup_directory_collection_links.'/cpuset/'.$Cpuset_path_job.'/memory.force_empty ]; then
                           echo 0 > '.$Cgroup_directory_collection_links.'/cpuset/'.$Cpuset_path_job.'/memory.force_empty
                         fi
-                        while ! oardodo rmdir '.$Cgroup_directory_collection_links.'/cpuset/'.$Cpuset_path_job.' ; do
-                          cat '.$Cgroup_directory_collection_links.'/cpuset/'.$Cpuset_path_job.'/tasks | xargs -n1 ps -fh -p 1>&2
-                          echo retry in 1s... 1>&2
-                          sleep 1
-                        done
-                        for d in '.$Cgroup_directory_collection_links.'/*/'.$Cpuset_path_job.'; do
-                          [ -w $d/memory.force_empty ] && echo 0 > $d/memory.force_empty
+                        for d in '.$Cgroup_directory_collection_links.'/cpuset/'.$Cpuset_path_job.'/* '.$Cgroup_directory_collection_links.'/cpuset/'.$Cpuset_path_job.'; do
                           if [ -d $d ]; then
-                            oardodo rmdir $d > /dev/null 2>&1 || exit 1
+                            [ -w $d/memory.force_empty ] && echo 0 > $d/memory.force_empty
+                            while ! oardodo rmdir $d ; do
+                              cat $d/tasks | xargs -n1 ps -fh -p 1>&2
+                              echo retry in 1s... 1>&2
+                              sleep 1
+                            done
+                          fi
+                        done
+                        for d in '.$Cgroup_directory_collection_links.'/*/'.$Cpuset_path_job.'/* '.$Cgroup_directory_collection_links.'/*/'.$Cpuset_path_job.'; do
+                          if [ -d $d ]; then
+                            [ -w $d/memory.force_empty ] && echo 0 > $d/memory.force_empty
+                            oardodo rmdir $d > /dev/null 2>&1
                           fi
                         done')){
                 # Uncomment this line if you want to use several network_address properties
